review changes: moved updating record of client registry to the epr functions.js
