+++ ui-sample/util-service/EPRFunctions.js
@@ -2,11 +2,20 @@ let Functions = require("./workflow/Functions");
 const _ = require('lodash')
 const async = require('async');
 const entityType = 'Employee';
+const appConfig = require('./sdk/appConfig');
+const RegistryService = require('./sdk/registryService')
+const KeycloakHelper = require('./sdk/KeycloakHelper');
+const httpUtils = require('./sdk/httpUtils.js');
+const logger = require('./sdk/log4j');
+const vars = require('./sdk/vars').getAllVars(process.env.NODE_ENV);
+const nerKeycloakHelper = new KeycloakHelper(vars.keycloak_ner);
+var registryService = new RegistryService();
 
 
 class EPRFunctions extends Functions {
     EPRFunctions() {
         setRequest(undefined)
+        setResponse(undefined)
     }
 
     getAdminUsers(callback) {
@@ -96,6 +105,90 @@ class EPRFunctions extends Functions {
                     callback(null, data)
                 });
                 break;
+        }
+    }
+
+    isIlimiUser(callback) {
+        let req = _.cloneDeep(this.request);
+        req.body.id = appConfig.APP_ID.READ;
+        registryService.readRecord(req, function (err, data) {
+            if (data && data.params.status === appConfig.STATUS.SUCCESSFULL) {
+                if (data.result[entityType].orgName === 'ILIMI')
+                    callback(null, data);
+                else
+                    callback(new Error("record does not belongs to the ILIMI org"))
+            } else
+                callback(err)
+        });
+    }
+
+    getNERToken(readResponse, callback) {
+        nerKeycloakHelper.getToken((err, token) => {
+            console.log("getNERToken", token)
+            if (token) callback(null, readResponse, token.access_token.token);
+            else callback(err)
+        });
+    }
+
+    getNERid(readResponse, token, callback) {
+        const headers = {
+            'content-type': 'application/json',
+            authorization: "Bearer " + token
+        }
+        const searchReq = {
+            body: {
+                id: appConfig.APP_ID.SEARCH,
+                request: {
+                    entityType: [entityType],
+                    filters: { email: { eq: readResponse.result[entityType].email } }
+                }
+            },
+            headers: headers,
+            url: vars.nerUtilServiceUrl + appConfig.UTILS_URL_CONFIG.SEARCH
+        }
+        httpUtils.post(searchReq, (err, res) => {
+            if (res.body.params.status === appConfig.STATUS.SUCCESSFULL && res.body.result[entityType].length > 0) {
+                callback(null, res.body, headers);
+            } else if (res.body.result[entityType].length <= 0) {
+                callback(new Error("error: record is not present in the client regitry"));
+            }
+        });
+    }
+
+    notifyNER(searchRes, headers, callback) {
+        let req = _.cloneDeep(this.request);
+        req.body.request[entityType].osid = searchRes.result[entityType][0].osid;
+        let option = {
+            body: updateReq.body,
+            headers: headers,
+            url: vars.nerUtilServiceUrl + appConfig.UTILS_URL_CONFIG.NOTIFICATIONS
+        }
+        httpUtils.post(option, (err, res) => {
+            if (res)
+                callback(null, res.body);
+            else
+                callback(err)
+        });
+    }
+
+    /**
+     * 
+     * @param {*} req 
+     */
+    updateRecordOfClientRegistry() {
+        if (this.response.data.params.status === appConfig.STATUS.SUCCESSFULL) {
+            logger.debug("updating record in client registry started", this.request.body)
+            async.waterfall([
+                this.isIlimiUser,
+                this.getNERToken,
+                this.getNERid,
+                this.notifyNER
+            ], (err, result) => {
+                if (result)
+                    logger.debug("Updating record in client registry is completed", result);
+                else
+                    logger.debug(err)
+            });
         }
     }
