+++ ui-sample/util-service/app.js
@@ -13,14 +13,9 @@ const templateConfig = require('./templates/template.config.json');
 const RegistryService = require('./sdk/registryService')
 const KeycloakHelper = require('./sdk/KeycloakHelper');
 const logger = require('./sdk/log4j');
-const httpUtils = require('./sdk/httpUtils.js');
-const vars = require('./sdk/vars').getAllVars(process.env.NODE_ENV);
-const appConfig = require('./sdk/appConfig');
 const port = process.env.PORT || 9081;
 let wfEngine = undefined
 const registryService = new RegistryService();
-const nerKeycloakHelper = new KeycloakHelper(vars.keycloak_ner);
-const entityType = 'Employee';
 
 app.use(cors())
 app.use(morgan('dev'));
@@ -31,16 +26,15 @@ const workFlowFunctionPre = (req) => {
     wfEngine.preInvoke(req);
 }
 
-const workFlowFunctionPost = (req) => {
-    wfEngine.postInvoke(req);
+const workFlowFunctionPost = (req, res) => {
+    wfEngine.postInvoke(req, res);
 }
 
 app.use((req, res, next) => {
     logger.info('pre api request interceptor');
     workFlowFunctionPre(req);
     next();
-    logger.info("post api request interceptor");
-    workFlowFunctionPost(req);
+    workFlowFunctionPost(req, res);
 });
 
 app.post("/registry/add", (req, res, next) => {
@@ -78,87 +72,12 @@ const getViewtemplate = (authToken) => {
 app.post("/registry/update", (req, res, next) => {
     registryService.updateRecord(req, function (err, data) {
         if (data) {
-            updateRecordOfClientRegistry(req, data);
             return res.send(data);
         } else {
             return res.send(err);
         }
     })
 });
-
-/**
- * 
- * @param {*} req 
- */
-const updateRecordOfClientRegistry = (req, res) => {
-    if (res.params.status === appConfig.STATUS.SUCCESSFULL) {
-        logger.debug("updating record in client registry started")
-        let updateReq = _.cloneDeep(req);
-        async.waterfall([
-            function (callback) {
-                req.body.id = appConfig.APP_ID.READ;
-                registryService.readRecord(req, function (err, data) {
-                    if (data && data.params.status === appConfig.STATUS.SUCCESSFULL) {
-                        if (data.result[entityType].orgName === 'ILIMI')
-                            callback(null, data);
-                        else
-                            callback("record does not belongs to the ILIMI org")
-                    } else
-                        callback(err)
-                });
-            },
-            function (readResponse, callback) {
-                nerKeycloakHelper.getToken((err, token) => {
-                    if (token) callback(null, readResponse, token.access_token.token);
-                    else callback(err)
-                });
-            },
-            function (readResponse, token, callback) {
-                const headers = {
-                    'content-type': 'application/json',
-                    authorization: "Bearer " + token
-                }
-                const searchReq = {
-                    body: {
-                        id: appConfig.APP_ID.SEARCH,
-                        request: {
-                            entityType: [entityType],
-                            filters: { email: { eq: readResponse.result[entityType].email } }
-                        }
-                    },
-                    headers: headers,
-                    url: vars.nerUtilServiceUrl + appConfig.UTILS_URL_CONFIG.SEARCH
-                }
-                httpUtils.post(searchReq, (err, res) => {
-                    if (res.body.params.status === appConfig.STATUS.SUCCESSFULL && res.body.result[entityType].length > 0) {
-                        callback(null, res.body, headers);
-                    } else if (res.body.result[entityType].length <= 0) {
-                        callback("error: record is not present in the client regitry");
-                    }
-                });
-            },
-            function (searchRes, headers, callback) {
-                updateReq.body.request[entityType].osid = searchRes.result[entityType][0].osid;
-                let option = {
-                    body: updateReq.body,
-                    headers: headers,
-                    url: vars.nerUtilServiceUrl + appConfig.UTILS_URL_CONFIG.NOTIFICATIONS
-                }
-                httpUtils.post(option, (err, res) => {
-                    if (res)
-                        callback(null, res.body);
-                    else
-                        callback(err)
-                });
-            }
-        ], function (err, result) {
-            if (result)
-                logger.debug("Updating record in client registry is completed", result);
-            else
-                logger.debug(err)
-        });
-    }
-}
 
 app.post("/notifications", (req, res, next) => {
     registryService.updateRecord(req, function (err, data) {
